chore(detector-alibaba): remove async detector
