--- detectors/node/opentelemetry-resource-detector-alibaba-cloud/src/detectors/constants.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-export const ALIBABA_CLOUD_IDMS_ENDPOINT = '100.100.100.200';
-export const ALIBABA_CLOUD_INSTANCE_IDENTITY_DOCUMENT_PATH =
-  '/latest/dynamic/instance-identity/document';
-export const ALIBABA_CLOUD_INSTANCE_HOST_DOCUMENT_PATH =
-  '/latest/meta-data/hostname';
-export const MILLISECONDS_TIME_OUT = 1000;
