+++ detectors/node/opentelemetry-resource-detector-alibaba-cloud/src/detectors/index.ts
@@ -15,4 +15,3 @@
  */
 
 export { alibabaCloudEcsDetector } from './AlibabaCloudEcsDetector';
-export { alibabaCloudEcsDetectorSync } from './AlibabaCloudEcsDetectorSync';
